utils.py разнесен на отдельные файлы
