+++ src/bot/handlers/about_fund_application.py
@@ -46,7 +46,7 @@ from bot.keyboards.about_fund_keyboards import (
     processes_anatomy_markup,
     things_path_markup,
 )
-from bot.utils import send_message
+from bot.utils.send_message import send_message
 
 
 async def handle_back_to_menu(
