+++ src/bot/handlers/contact_list_application.py
@@ -13,7 +13,7 @@ from bot.constants.state import CONTACT_LIST
 from bot.handlers.command_application import stop_callback
 from bot.keyboards.basic_info_keyboards import contact_list_markup
 from bot.keyboards.keyboards import main_menu_markup
-from bot.utils import find_contacts
+from bot.utils.keyword_search import find_contacts
 
 
 async def contact_list_callback(
@@ -38,9 +38,7 @@ async def find_contact_by_fio(
     и возвращает контакт (телефон, имя...).'''
     user_text = update.message.text
     answer = await find_contacts(user_text)
-    await update.message.reply_text(
-        text=answer, reply_markup=main_menu_markup
-    )
+    await update.message.reply_text(text=answer, reply_markup=main_menu_markup)
     return ConversationHandler.END
